+++ meetup2xibo/log_summarizer/command_line.py
@@ -11,26 +11,34 @@ parser.add_argument(
         nargs='?',
         type=argparse.FileType('r'),
         default=sys.stdin,
-        help="Input file path [default: standard input]")
+        help="Input file path (default: standard input)")
 
 parser.add_argument(
         'outfile',
         nargs='?',
         type=argparse.FileType('w'),
         default=sys.stdout,
-        help="Output file path [default: standard output]")
+        help="Output file path (default: standard output)")
+
+parser.add_argument(
+        '-m', '--mappings',
+        action='store_true',
+        help='Summarize location mappings in CSV format. '
+             '(default: summarize logs in HTML format)')
 
 parser.add_argument(
         '-s', '--subject',
         dest='email_subject',
-        default="",
-        help='Generate email headers including this Subject.')
+        default="Meetup to Xibo log summary",
+        help='Email subject. (default: %(default)s)')
 
 parser.add_argument(
         '-t', '--to',
         dest='email_to',
         default="",
-        help='Generate email headers including this To address.')
+        help='Generate an email message to this address or space '
+             'separated addressses. Overrides --mappings option. '
+             '(default: no email message)')
 
 
 def parse_args(args=None):