+++ meetup2xibo/updater/meetup_api.py
@@ -7,7 +7,8 @@ import requests
 
 class MeetupEventsRetriever:
 
-    def __init__(self, group_url_name, api_key, events_wanted, cancelled_last_time):
+    def __init__(
+            self, group_url_name, api_key, events_wanted, cancelled_last_time):
         """Initialize with a Meetup group URL name, a Meetup API, the number of
         events wanted from Meetup, and the last time allowed for cancelled
         events."""