+++ tests/updater/conftest.py
@@ -1,6 +1,5 @@
 """Test generating the Xibo API."""
 
-from ..context import meetup2xibo
 from meetup2xibo.updater.xibo_api_url_builder import XiboApiUrlBuilder
 from meetup2xibo.updater.site_cert_assurer import assure_site_cert
 from meetup2xibo.updater.oauth2_session_starter import Oauth2SessionStarter