+++ tests/updater/test_anti_flapper.py
@@ -1,8 +1,8 @@
 """Test the anti-flapper's judgements about event times."""
 
-from ..context import meetup2xibo
 from meetup2xibo.updater.xibo_event import XiboEvent
-from meetup2xibo.updater.anti_flapper import AntiFlapper, iso_offset_time
+from meetup2xibo.updater.anti_flapper import AntiFlapper, EventFlappingStatus, \
+    iso_offset_time
 from hypothesis import given, assume, example
 import hypothesis.strategies as st
 from datetime import datetime
@@ -29,66 +29,82 @@ def make_xibo_event(start_time, end_time):
         end_time = end_time
         )
 
-def assert_event_ok(event_start, event_end, recent, current, future):
-    """Assert that an event with start and end times falls outside the flapping windows."""
+def assert_event_status(status, event_start, event_end, recent, current, future):
+    """Assert that an event with start and end times has the expected status
+    relative to the flapping windows."""
     anti_flapper = AntiFlapper(recent, current, future)
     xibo_event = make_xibo_event(event_start, event_end)
-    assert anti_flapper.is_ok(xibo_event)
+    assert anti_flapper.categorize(xibo_event) == status
 
-def assert_event_not_ok(event_start, event_end, recent, current, future):
-    """Assert that an event with start and end times does not fall outside the
+def assert_event_retire(event_start, event_end, recent, current, future):
+    """Assert retiring an event with start and end times relative to the
     flapping windows."""
-    anti_flapper = AntiFlapper(recent, current, future)
-    xibo_event = make_xibo_event(event_start, event_end)
-    assert not anti_flapper.is_ok(xibo_event)
+    status = EventFlappingStatus.retire
+    assert_event_status(status, event_start, event_end, recent, current, future)
+
+def assert_event_delete(event_start, event_end, recent, current, future):
+    """Assert deleting an event with start and end times relative to the
+    flapping windows."""
+    status = EventFlappingStatus.delete
+    assert_event_status(status, event_start, event_end, recent, current, future)
+
+def assert_event_keep(event_start, event_end, recent, current, future):
+    """Assert keeping an event with start and end times relative to the
+    flapping windows."""
+    status = EventFlappingStatus.keep
+    assert_event_status(status, event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_past_event_ok(iso_dates):
-    """Test that an event ending before flapping windows is ok."""
+def test_past_event_retire(iso_dates):
+    """Test that an event ending before should be retired."""
     event_start, event_end, recent, current, future = tuple(iso_dates)
-    assert_event_ok(event_start, event_end, recent, current, future)
+    assert_event_retire(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_recent_event_ok(iso_dates):
-    """Test that an event ending during the recent to current flapping windows is not ok."""
+def test_recent_event_keep(iso_dates):
+    """Test that an event ending during the recent to current flapping window
+    should be kept."""
     event_start, recent, event_end, current, future = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_current_event_ok(iso_dates):
-    """Test that an event within the recent to current flapping windows is not ok."""
+def test_current_event_keep(iso_dates):
+    """Test that an event within the recent to current flapping windows should
+    be kept."""
     recent, event_start, event_end, current, future = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_ongoing_event_ok(iso_dates):
-    """Test that an event starting within the recent to current flapping windows is not ok."""
+def test_ongoing_event_keep(iso_dates):
+    """Test that an event starting within the recent to current flapping
+    windows should be kept."""
     recent, event_start, current, event_end, future = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_surrounding_event_ok(iso_dates):
-    """Test that an event surrounding the recent to current flapping windows is not ok."""
+def test_surrounding_event_keep(iso_dates):
+    """Test that an event surrounding the recent to current flapping windows
+    should be kept."""
     event_start, recent, current, event_end, future = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_planned_event_ok(iso_dates):
-    """Test that an event planned between the flapping windows is ok."""
+def test_planned_event_delete(iso_dates):
+    """Test that an event planned between the flapping windows should be deleted."""
     recent, current, event_start, event_end, future = tuple(iso_dates)
-    assert_event_ok(event_start, event_end, recent, current, future)
+    assert_event_delete(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_future_event_ok(iso_dates):
-    """Test that a future event ending after the future time is not ok."""
+def test_future_event_keep(iso_dates):
+    """Test that a future event ending after the future time should be kept."""
     recent, current, event_start, future, event_end = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 @given(iso_dates = iso_date_lists)
-def test_far_future_event_ok(iso_dates):
-    """Test that a future event starting after the future time is not ok."""
+def test_far_future_event_keep(iso_dates):
+    """Test that a future event starting after the future time should be kept."""
     recent, current, future, event_start, event_end = tuple(iso_dates)
-    assert_event_not_ok(event_start, event_end, recent, current, future)
+    assert_event_keep(event_start, event_end, recent, current, future)
 
 def test_iso_offset_time_future():
     """Test offsetting time into the future."""