+++ tests/updater/test_event_converter.py
@@ -1,6 +1,5 @@
 """Test the event converter from Meetup JSON to event object."""
 
-from ..context import meetup2xibo
 from meetup2xibo.updater.event_converter import EventConverter, Event, PartialEvent
 from meetup2xibo.updater.location_chooser import LocationChooser
 from hypothesis import given, assume, example